+++ CHANGELOG.md
@@ -52,6 +52,12 @@
 
  - Correctly load saved value for mm options See [#1136](https://github.com/freesewing/freesewing/issues/1136)
 
+### core
+
+#### Fixed
+
+ - Fix a bug in `path.shiftAlong` where no point is returned if the distance to shift is a fraction of one step (1/25mm) into a new path segment See [#1140](https://github.com/freesewing/freesewing/issues/1140)
+
 ### diana
 
 #### Changed